libs/requirements.txt
pyspark==3.1.*
pandas==1.2.*
delta-spark==1.0.0
pytest
pytest-azurepipelines
pytest-cov
